src/python/requirements.txt
textacy==0.6.2
spacy==2.1.3
msgpack==0.5.6
msgpack-numpy==0.4.3.2
gensim==3.6.0
zensols.actioncli>=1.0.15
